[rdar://116400979] Refactor DocumentationTopic view to be able to reuse the navigator for other views
