+++ src/components/Navigator/NavigatorDataProvider.vue
@@ -1,7 +1,7 @@
 <!--
   This source file is part of the Swift.org open source project
 
-  Copyright (c) 2022 Apple Inc. and the Swift project authors
+  Copyright (c) 2022-2024 Apple Inc. and the Swift project authors
   Licensed under Apache License v2.0 with Runtime Library Exception
 
   See https://swift.org/LICENSE.txt for license information
@@ -31,7 +31,7 @@ export default {
      */
     technologyUrl: {
       type: String,
-      required: true,
+      default: '',
     },
     apiChangesVersion: {
       type: String,
