+++ src/styles/core/_breakpoints.scss
@@ -1,7 +1,7 @@
 /**
  * This source file is part of the Swift.org open source project
  *
- * Copyright (c) 2021-2023 Apple Inc. and the Swift project authors
+ * Copyright (c) 2021-2024 Apple Inc. and the Swift project authors
  * Licensed under Apache License v2.0 with Runtime Library Exception
  *
  * See https://swift.org/LICENSE.txt for license information
@@ -16,7 +16,8 @@
 $breakpoint-attributes: (
   default: (
     xlarge: (
-      min-width: 1920px
+      min-width: 1920px,
+      content-width: 1536px,
     ),
     large: (
       min-width: 1251px,
