+++ tests/unit/components/BreakpointEmitter.spec.js
@@ -1,7 +1,7 @@
 /**
  * This source file is part of the Swift.org open source project
  *
- * Copyright (c) 2021 Apple Inc. and the Swift project authors
+ * Copyright (c) 2021-2024 Apple Inc. and the Swift project authors
  * Licensed under Apache License v2.0 with Runtime Library Exception
  *
  * See https://swift.org/LICENSE.txt for license information
@@ -125,7 +125,7 @@ describe('BreakpointEmitter', () => {
       removeListener,
     }));
     shallowMount(BreakpointEmitter, { scopedSlots }).destroy();
-    expect(removeListener).toBeCalledTimes(3);
+    expect(removeListener).toBeCalledTimes(4);
   });
 
   it('supports scopes', async () => {
