+++ tests/unit/components/Pager.spec.js
@@ -9,6 +9,9 @@
 */
 import { shallowMount } from '@vue/test-utils';
 import Pager from 'docc-render/components/Pager.vue';
+import { BreakpointAttributes } from 'docc-render/utils/breakpoints';
+
+const GUTTERS_WIDTH = 174;
 
 describe('Pager', () => {
   const propsData = {
@@ -41,6 +44,63 @@ describe('Pager', () => {
 
     const indicators = wrapper.findAll('.indicator');
     expect(indicators.length).toBe(propsData.pages.length);
+  });
+
+  it('collapses the controllers if contentWidth is smaller than the large contentWidth + the gutters width in large viewport', () => {
+    // set large viewport
+    window.innerWidth = BreakpointAttributes.default.large.minWidth + 1;
+    const wrapper = shallowMount(Pager, { propsData });
+    wrapper.setData({
+      appState: {
+        contentWidth: BreakpointAttributes.default.large.contentWidth + GUTTERS_WIDTH + 10,
+      },
+    });
+
+    expect(wrapper.classes('with-compact-controls')).toBe(false);
+
+    wrapper.setData({
+      appState: {
+        contentWidth: BreakpointAttributes.default.large.contentWidth + GUTTERS_WIDTH - 10,
+      },
+    });
+
+    expect(wrapper.classes('with-compact-controls')).toBe(true);
+  });
+
+  it('collapses the controllers if contentWidth is smaller than the medium contentWidth + the gutters width in medium viewport', () => {
+    // set medium viewport
+    window.innerWidth = BreakpointAttributes.default.medium.maxWidth - 1;
+    const wrapper = shallowMount(Pager, { propsData });
+
+    wrapper.setData({
+      appState: {
+        contentWidth: BreakpointAttributes.default.medium.contentWidth + GUTTERS_WIDTH + 10,
+      },
+    });
+
+    expect(wrapper.classes('with-compact-controls')).toBe(false);
+
+    wrapper.setData({
+      appState: {
+        contentWidth: BreakpointAttributes.default.medium.contentWidth + GUTTERS_WIDTH - 10,
+      },
+    });
+
+    expect(wrapper.classes('with-compact-controls')).toBe(true);
+  });
+
+  it('collapses the controllers in small viewports', () => {
+    // set small viewport
+    window.innerWidth = BreakpointAttributes.default.small.minWidth;
+    const wrapper = shallowMount(Pager, { propsData });
+
+    wrapper.setData({
+      appState: {
+        contentWidth: BreakpointAttributes.default.small.minWidth,
+      },
+    });
+
+    expect(wrapper.classes('with-compact-controls')).toBe(true);
   });
 
   it('renders each page using provided slots', () => {
